requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pydantic==2.5.0
slowapi==0.1.8
redis==5.0.1
chromadb==0.4.22
aiohttp==3.9.1
openai==1.6.0
python-multipart
prometheus-fastapi-instrumentator
prometheus-client
python-json-logger
